add tests

+++ src/main/java/com/circleescape/server/scoreboard/data/ScoreRepository.java
@@ -2,5 +2,5 @@ package com.circleescape.server.scoreboard.data;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ScoreRepository extends CrudRepository<Score, Integer> {
+public interface ScoreRepository extends CrudRepository<Score, Long> {
 }
